+++ pkg/deployment/reconcile/plan_builder_high.go
@@ -55,6 +55,7 @@ func (r *Reconciler) createHighPlan(ctx context.Context, apiObject k8sutil.APIOb
 		ApplyIfEmpty(r.updateMemberUpdateConditionsPlan).
 		ApplyIfEmpty(r.updateMemberRotationConditionsPlan).
 		ApplyIfEmpty(r.createMemberAllowUpgradeConditionPlan).
+		ApplyIfEmpty(r.createMemberGatewayConfigConditionPlan).
 		ApplyIfEmpty(r.createMemberRecreationConditionsPlan).
 		ApplyIfEmpty(r.createMemberPodSchedulingFailurePlan).
 		ApplyIfEmpty(r.createRotateServerStoragePVCPendingResizeConditionPlan).