+++ pkg/util/grpc/grpc.go
@@ -156,7 +156,11 @@ type GRPC[T proto.Message] struct {
 	Object T
 }
 
-func (g *GRPC[T]) UnmarshalJSON(data []byte, opts ...util.Mod[protojson.UnmarshalOptions]) error {
+func (g *GRPC[T]) UnmarshalJSON(data []byte) error {
+	return g.UnmarshalJSONOpts(data)
+}
+
+func (g *GRPC[T]) UnmarshalJSONOpts(data []byte, opts ...util.Mod[protojson.UnmarshalOptions]) error {
 	o, err := Unmarshal[T](data, opts...)
 	if err != nil {
 		return err
@@ -166,6 +170,10 @@ func (g *GRPC[T]) UnmarshalJSON(data []byte, opts ...util.Mod[protojson.Unmarsha
 	return nil
 }
 
-func (g GRPC[T]) MarshalJSON(opts ...util.Mod[protojson.MarshalOptions]) ([]byte, error) {
+func (g GRPC[T]) MarshalJSON() ([]byte, error) {
+	return g.MarshalJSONOpts()
+}
+
+func (g GRPC[T]) MarshalJSONOpts(opts ...util.Mod[protojson.MarshalOptions]) ([]byte, error) {
 	return Marshal[T](g.Object, opts...)
 }